Add meal start day to User
In preperation for the planner page, added a new field to the User
Entity 'MealStartDay' which is a day of the week.

MealStartDay will determine the week view the user will see, for example
if they select 'Wednesday' their planner view will start on Wednesday
and end on Tuesday.

As a side, the name and email fields in the update form are no longer
optional (pointers), as they are always provided on updates (only
password is now optional), doing this has simplified the update account
code.