Move meal handlers to web package
It has become necessary to move the meal handlers to the web package.
This is because the new GetPlannerHandler has a reliance on the Auth
package (to fetch the user). Moving them to web also makes sense as that
is where the template files live, they are specifically for the web app
so it feels like a better fit.

The auth handlers will be moved to web soon.